Renamed mpc-pointing into deburring_mpc
