+++ README.md
@@ -4,6 +4,6 @@ Set of tools to work on manipulation tasks with the robot Talos.
 
 This repo contains several package:
 
-- mpc-pointing: an implementation of a Model Predictive Controller based on Crocoddyl
+- deburring_mpc: an implementation of a Model Predictive Controller based on Crocoddyl
 - ros-interface: the interface to allow the use of the MPC within a ROS architecture
 - python: a python example of the use of the MPC with the PyBullet simulator
