+++ deburring-mpc/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>mpc-pointing</name>
+  <name>deburring_mpc</name>
   <version>0.0.0</version>
   <description>MPC for manipulation tasks with Talos</description>
   <maintainer email="come.perrot@laas.fr">Côme Perrot</maintainer>
