+++ deburring-mpc/python/main.cpp
@@ -1,8 +1,8 @@
 #include <eigenpy/eigenpy.hpp>
 
-#include "mpc-pointing/python-fwd.hpp"
+#include "deburring_mpc/python-fwd.hpp"
 
-BOOST_PYTHON_MODULE(mpc_pointing_pywrap) {
+BOOST_PYTHON_MODULE(deburring_mpc_pywrap) {
   boost::python::import("pinocchio");
   boost::python::import("crocoddyl");
   // Enabling eigenpy support, i.e. numpy/eigen compatibility.
