+++ deburring-mpc/src/mpc-debug.cpp
@@ -1,6 +1,6 @@
 #include <yaml-cpp/yaml.h>
 
-#include "mpc-pointing/mpc.hpp"
+#include "deburring_mpc/mpc.hpp"
 
 namespace mpc_p {
