+++ deburring-mpc/src/ocp-cost-library.cpp
@@ -1,4 +1,4 @@
-#include "mpc-pointing/ocp.hpp"
+#include "deburring_mpc/ocp.hpp"
 
 namespace mpc_p {
 void OCP_Point::defineFeetContact(Contact &contactCollector) {
