+++ deburring-mpc/src/ocp-params.cpp
@@ -1,6 +1,6 @@
 #include <yaml-cpp/yaml.h>
 
-#include "mpc-pointing/ocp.hpp"
+#include "deburring_mpc/ocp.hpp"
 
 namespace mpc_p {
