+++ deburring-mpc/src/ocp-problem-helper.cpp
@@ -1,4 +1,4 @@
-#include "mpc-pointing/ocp.hpp"
+#include "deburring_mpc/ocp.hpp"
 
 namespace mpc_p {
 // Functions to interact with ddp
