+++ deburring-mpc/tests/test_OCP.cpp
@@ -1,4 +1,4 @@
-#include "mpc-pointing/ocp.hpp"
+#include "deburring_mpc/ocp.hpp"
 
 #define PI 3.14159265
