+++ ros-interface/package.xml
@@ -15,7 +15,7 @@
   <depend>eigen_conversions</depend>
   <depend>tf2_ros</depend>
   <depend>linear_feedback_controller_msgs</depend>
-  <depend>mpc-pointing</depend>
+  <depend>deburring_mpc</depend>
   <depend>pal_statistics</depend>
 
   <build_depend>rostest</build_depend>
