+++ ros-interface/src/ros-mpc-pointing.cpp
@@ -1,4 +1,4 @@
-#include <mpc-pointing/mpc.hpp>
+#include <deburring_mpc/mpc.hpp>
 // Must be included first
 
 #include <pal_statistics/pal_statistics_macros.h>
